+++ lib/src/json/room_object_ambiance.dart
@@ -1,7 +1,6 @@
 import 'dart:math';
 
-import 'package:angstrom/angstrom.dart' show RoomObject;
-import 'package:angstrom/src/room_object.dart' show RoomObject;
+import 'package:angstrom/angstrom.dart' show RoomObject, SoundReference;
 import 'package:json_annotation/json_annotation.dart';
 
 part 'room_object_ambiance.g.dart';
@@ -11,10 +10,10 @@ part 'room_object_ambiance.g.dart';
 class RoomObjectAmbiance {
   /// Create an instance.
   const RoomObjectAmbiance({
-    required this.path,
+    required this.id,
+    required this.soundReference,
     required this.x,
     required this.y,
-    required this.volume,
     required this.ambianceMaxDistance,
   });
 
@@ -22,11 +21,11 @@ class RoomObjectAmbiance {
   factory RoomObjectAmbiance.fromJson(final Map<String, dynamic> json) =>
       _$RoomObjectAmbianceFromJson(json);
 
-  /// The path to the sound to play.
-  final String path;
+  /// The ID of this object whose ambiance this instance represents.
+  final String id;
 
-  /// The volume of the sound.
-  final double volume;
+  /// The sound reference to play.
+  final SoundReference soundReference;
 
   /// The maximum distance at which this ambiance will be heard.
   final int ambianceMaxDistance;